% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_ini}
\alias{read_ini}
\title{Read Configuration Data from an INI File}
\usage{
read_ini(filename)
}
\arguments{
\item{filename}{A string specifying the path to the input INI file.}
}
\value{
A named list where each element is itself a named list representing sections
and key-value pairs of the configuration data.
}
\description{
This function reads an INI file and converts its contents into an R list structure.
}
\details{
The \code{read_ini} function parses the specified INI file and converts its sections
and key-value pairs into a nested R list. Each section in the INI file becomes
a named list within the result, and each key-value pair within a section becomes
an element of that list.

If the INI file cannot be loaded, the function will raise an error.
}
\examples{
config <- list(
    section1 = list(
      key1 = "value1"
    ),
    section2 = list(
      keyA = "valueA",
      keyB = "valueB"
    )
)
 ini_file <- tempfile()
 write_ini(config, ini_file)
 read_ini(ini_file)

 unlink(ini_file)

}
