% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_ini}
\alias{write_ini}
\title{Write Configuration Data to an INI File}
\usage{
write_ini(config_list, filename)
}
\arguments{
\item{config_list}{A named list where each element is itself a named list representing sections and key-value pairs of the configuration data.}

\item{filename}{A string specifying the path to the output INI file.}
}
\value{
This function does not return a value. It writes the INI file to the specified path.
}
\description{
This function converts an R list structure into an INI file format and writes it to a specified file.
}
\details{
The \code{config_list} parameter should be a named list where each top-level name represents a section in the INI file.
Each section should be a named list where the names are the keys and the values are the corresponding values in the INI file.
Numeric values are converted to strings before being written to the file.

If the \code{config_list} is empty, an empty INI file will be created.
}
\examples{
config <- list(
  section1 = list(
    key1 = "value1",
    key2 = 123
  ),
  section2 = list(
    keyA = "valueA",
    keyB = "valueB"
  )
)
ini_file <- tempfile()
write_ini(config, ini_file)
unlink(ini_file)

}
